fix(conf): don't use `cache` dir as base for all data dirs
This is a bit tricky but the issue is that in a `for` loop, the loop
variable(s) are shared during the whole iteration. So capturing its
value in the `lambda` as `k=k` doesn't capture the value of the current
iteration, but the value of the last one.
In this code that happened to be `cache`, so all `data_dirs` usage was
ending up with a path under `directory_cache`.